requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.0
requests==2.32.3
pydantic==2.10.3
transformers==4.47.1
torch==2.5.1
sentencepiece==0.2.0